planting month displays the correct info
