+++ seeds/garden.js
@@ -15,12 +15,7 @@ exports.seed = function(knex, Promise) {
         {veg_id: 8, user_id: 1},
         {veg_id: 1, user_id: 2},
         {veg_id: 2, user_id: 2},
-        {veg_id: 3, user_id: 2},
-        {veg_id: 4, user_id: 2},
-        {veg_id: 5, user_id: 2},
-        {veg_id: 6, user_id: 2},
-        {veg_id: 7, user_id: 2},
-        {veg_id: 8, user_id: 2},
+        {veg_id: 3, user_id: 2}
       ]);
     });
 };
