+++ server/routes/users.js
@@ -12,10 +12,9 @@ router.post('/', (req,res)=>{
 })
 
 router.post('/info', (req,res)=>{
-  console.log('routes', req.body)
-  db.addToGarden(req.body.veg, req.body.user)
-  .then(user =>{
-    res.json(user)
+  db.addToGarden(req.body.veg, req.body.user[0])
+  .then(veg =>{
+    res.json(veg)
   })
 })
